LETSGOOOOOO

+++ setup.sh
@@ -15,18 +15,22 @@ sed -i "s/backend.example.com/${BACK_DNS}/g" webconf/sites-enabled/backend
 sed -i "s/frontend.example.com/${FRONT_DNS}/g" webconf/sites-enabled/frontend
 
 if [ ! -f /etc/letsencrypt/live/${BACK_DNS}/fullchain.pem ]; then
-    echo "Please make sure the domain cert is valid"
+    printf "\033[0;31mPlease make sure the domain cert is valid\033[0m\n"
     exit 1
 fi
 if [ ! -f /etc/letsencrypt/live/${BACK_DNS}/privkey.pem ]; then
-    echo "Please make sure the domain cert is valid"
+    printf "\033[0;31mPlease make sure the domain cert is valid\033[0m\n"
     exit 1
 fi
 
 sudo docker compose pull
-# check docker compose is running
-# use docker compose ps to check
+# check docker container is already running
+# if runninng, compose down
+# then compose up
+if [ "$(sudo docker ps -q -f name=taxi-server)" ]; then
+    sudo docker compose down
+fi
 
-
-sudo docker compose down
 sudo docker compose up -d
+
+printf "\033[0;32mDeploy Success!\033[0m\n"
